requirements.txt
numpy==1.11.3
appdirs==1.4.3
seaborn==0.7.1
pandas==0.19.2
matplotlib==2.0.0
keras>1.2.0
bokeh==0.12.6
scipy==0.18.1
h5py==2.6.0
tensorflow-gpu>=1.5.0
scikit-image==0.13.0
opencv-python==3.2.0.7
Pillow==4.0.0
PyYAML==3.12